Move fixtures into folder
