+++ src/core/source/MemorySource.test.ts
@@ -1,6 +1,6 @@
 import fs from 'node:fs/promises'
 import {suite} from '@alinea/suite'
-import demoTree from '../../test/demo.json' with {type: 'json'}
+import demoTree from '../../test/fixtures/demo.json' with {type: 'json'}
 import {ReadonlyTree} from '../source/Tree.js'
 import {MemorySource} from './MemorySource.js'
