+++ src/core/source/Tree.test.ts
@@ -1,5 +1,7 @@
 import {suite} from '@alinea/suite'
-import treeExample from '../../test/exampleTree.json' with {type: 'json'}
+import treeExample from '../../test/fixtures/exampleTree.json' with {
+  type: 'json'
+}
 import {ReadonlyTree, WriteableTree} from './Tree.js'
 
 const test = suite(import.meta)
